+++ apps/web/components/Feature/Auth/SignUpViewer/index.tsx
@@ -1,31 +1,23 @@
 "use client";
 
-import { useMemo, useState } from "react";
 import Image from "@/components/UI/SafeImage";
-import { useRouter } from "next/navigation";
 import { useTranslation } from "react-i18next";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { FormProvider, useForm, useWatch } from "react-hook-form";
+import { FormProvider, useWatch } from "react-hook-form";
 import { EyeClosedIcon, EyeOpenIcon } from "@/assets/icons";
 import logo from "@/assets/icons/Kiibee_logo_mark_black.svg";
 import AuthBackButton from "@/components/Feature/Auth/AuthBackButton";
 import GenericButton from "@/components/UI/GenericButton";
 import FormField from "@/components/UI/FormField";
 import { MonoText } from "@/components/UI/Monotext";
-import { useViewerSignUp } from "@/hooks/auth/useViewerSignUp";
-import { persistAuthSession } from "@/lib/auth/authSession";
-import { useApiErrorMessage } from "@/lib/http/useApiErrorMessage";
 import { ALERT } from "@/utils/common";
 import { PATHS } from "@/utils/path";
 import { INPUT_TYPE } from "@/utils/ui";
-import { createViewerSignupSchema } from "@/lib/validation/schema";
 import {
   PASSWORD_FIELD_KEYS,
   PasswordVisibility,
   VIEWER_FIELDS,
   ViewerFieldConfig,
   ViewerFieldKey,
-  ViewerFormValues,
 } from "@/utils/signup";
 import {
   Card,
@@ -41,6 +33,7 @@ import {
   Title,
   Wrapper,
 } from "./styles";
+import { useViewerSignUpForm } from "@/hooks/Auth/useViewerSignUpForm";
 
 type PasswordFieldKey = keyof PasswordVisibility;
 
@@ -50,85 +43,25 @@ const isPasswordField = (
   PASSWORD_FIELD_KEYS.includes(fieldKey as PasswordFieldKey);
 
 export default function SignUpViewer() {
-  const router = useRouter();
   const { t } = useTranslation();
-  const { mutateAsync: viewerSignUp, isPending: isSubmitting } =
-    useViewerSignUp();
-  const [passwordVisibility, setPasswordVisibility] =
-    useState<PasswordVisibility>({
-      password: false,
-      repeatPassword: false,
-    });
-  const [formError, setFormError] = useState("");
-  const { getErrorMessage, applyFieldErrors } = useApiErrorMessage();
-  const schema = useMemo(
-    () =>
-      createViewerSignupSchema({
-        fullNameRequired: t("viewerSignup.form.fixHighlightedFields"),
-        emailRequired: t("authForm.errors.emailRequired"),
-        emailInvalid: t("authForm.errors.emailInvalid"),
-        passwordRequired: t("authForm.errors.passwordRequired"),
-        repeatPasswordRequired: t("viewerSignup.form.fixHighlightedFields"),
-        passwordMismatch: t("viewerSignup.form.passwordMismatch"),
-        consentRequired: t("viewerSignup.form.fixHighlightedFields"),
-      }),
-    [t],
-  );
-  type ViewerSignupValues = ReturnType<typeof schema.parse>;
-  const methods = useForm<ViewerSignupValues>({
-    resolver: zodResolver(schema),
-    mode: "onChange",
-    defaultValues: {
-      fullName: "",
-      email: "",
-      password: "",
-      repeatPassword: "",
-      agreed: false,
-    },
-  });
+
   const {
+    methods,
+    isValid,
+    errors,
+    isSubmitting,
+    formError,
+    passwordVisibility,
+    updateField,
+    togglePassword,
     handleSubmit,
-    setValue,
-    setError,
-    formState: { isValid, errors },
-  } = methods;
+  } = useViewerSignUpForm();
+
   const agreedValue = useWatch({
     control: methods.control,
     name: "agreed",
     defaultValue: false,
   });
-
-  const onSubmit = async (values: ViewerSignupValues) => {
-    setFormError("");
-    try {
-      const response = await viewerSignUp({
-        fullName: values.fullName.trim(),
-        email: values.email.trim(),
-        password: values.password,
-        confirmPassword: values.repeatPassword,
-      });
-
-      persistAuthSession(response);
-
-      router.push("/auth/signup-viewer/preferences");
-    } catch (error) {
-      applyFieldErrors(error, setError);
-      setFormError(getErrorMessage(error, "viewerSignup.form.signupFailed"));
-    }
-  };
-
-  const updateField = (
-    field: keyof ViewerFormValues,
-    value: string | boolean,
-  ) => {
-    setValue(field, value as never, {
-      shouldDirty: true,
-      shouldValidate: true,
-    });
-    if (formError) {
-      setFormError("");
-    }
-  };
 
   const renderField = (field: ViewerFieldConfig) => {
     const passwordFieldKey = isPasswordField(field.key) ? field.key : null;
@@ -142,7 +75,7 @@ export default function SignUpViewer() {
       : field.type;
 
     return (
-      <FormField<ViewerSignupValues>
+      <FormField
         key={field.key}
         id={`viewer-${field.key}`}
         name={field.key}
@@ -161,13 +94,7 @@ export default function SignUpViewer() {
           ) : undefined
         }
         onIconClick={
-          passwordFieldKey
-            ? () =>
-                setPasswordVisibility((prev) => ({
-                  ...prev,
-                  [passwordFieldKey]: !prev[passwordFieldKey],
-                }))
-            : undefined
+          passwordFieldKey ? () => togglePassword(passwordFieldKey) : undefined
         }
         required
       />
@@ -185,7 +112,7 @@ export default function SignUpViewer() {
           </Title>
 
           <FormProvider {...methods}>
-            <Form onSubmit={handleSubmit(onSubmit)}>
+            <Form onSubmit={handleSubmit}>
               {VIEWER_FIELDS.map((field) => renderField(field))}
 
               <CheckboxRow>
