+++ apps/web/hooks/auth/useCreatorRequestForm.ts
@@ -1,107 +1,139 @@
 "use client";
 
 import { useMemo, useState } from "react";
+import { useTranslation } from "react-i18next";
 import { useRouter } from "next/navigation";
 import { toast } from "react-toastify";
-import {
-  CreatorFormValues,
-  INITIAL_CREATOR_FORM,
-  REQUIRED_CREATOR_FIELD_KEYS,
-} from "@/utils/authCreatorForm";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import { useApiErrorMessage } from "@/lib/http/useApiErrorMessage";
+import { createCreatorRequestSchema } from "@/lib/validation/schema";
 import { PATHS } from "@/utils/path";
 import { useCreatorRequest } from "./useCreatorRequest";
-import { FORM_MESSAGE_TONE, FormMessageTone } from "@/utils/ui";
 
-export function useCreatorRequestForm(t: (key: string) => string) {
+export type CreatorRequestValues = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  phone: string;
+  cvr: string;
+  address: string;
+  city: string;
+  postalCode: string;
+  workLink: string;
+  contentDescription: string;
+  agreed: boolean;
+};
+
+export function useCreatorRequestForm() {
+  const { t } = useTranslation();
   const router = useRouter();
-  const creatorRequest = useCreatorRequest();
-
-  const [formValues, setFormValues] =
-    useState<CreatorFormValues>(INITIAL_CREATOR_FORM);
 
   const [formMessage, setFormMessage] = useState("");
-  const [messageTone, setMessageTone] = useState<FormMessageTone>(
-    FORM_MESSAGE_TONE.ERROR,
-  );
-  const isSubmitEnabled = useMemo(() => {
-    const hasRequiredValues = REQUIRED_CREATOR_FIELD_KEYS.every((key) =>
-      Boolean(formValues[key].trim()),
-    );
+  const [messageTone, setMessageTone] = useState<"error" | "success">("error");
 
-    return hasRequiredValues && formValues.agreed && !creatorRequest.isPending;
-  }, [creatorRequest.isPending, formValues]);
+  const creatorRequest = useCreatorRequest();
+  const { getErrorMessage, applyFieldErrors } = useApiErrorMessage();
+
+  const schema = useMemo(
+    () =>
+      createCreatorRequestSchema({
+        firstNameRequired: t("authCreator.form.firstName"),
+        lastNameRequired: t("authCreator.form.lastName"),
+        emailRequired: t("authForm.errors.emailRequired"),
+        emailInvalid: t("authForm.errors.emailInvalid"),
+        addressRequired: t("authCreator.form.address"),
+        cityRequired: t("authCreator.form.city"),
+        postalCodeRequired: t("authCreator.form.postalCode"),
+        workLinkRequired: t("authCreator.form.workLink"),
+        workLinkInvalid: t("authCreator.form.workLinkInvalid"),
+        contentDescriptionRequired: t("authCreator.form.contentLabel"),
+        consentRequired: t("authCreator.form.consentPrefix"),
+      }),
+    [t],
+  );
+
+  const methods = useForm<CreatorRequestValues>({
+    resolver: zodResolver(schema),
+    mode: "onChange",
+    defaultValues: {
+      firstName: "",
+      lastName: "",
+      email: "",
+      phone: "",
+      cvr: "",
+      address: "",
+      city: "",
+      postalCode: "",
+      workLink: "",
+      contentDescription: "",
+      agreed: false,
+    },
+  });
+
+  const {
+    handleSubmit,
+    setValue,
+    reset,
+    setError,
+    formState: { isValid, errors },
+  } = methods;
 
   const updateField = (
-    field: keyof CreatorFormValues,
+    field: keyof CreatorRequestValues,
     value: string | boolean,
   ) => {
-    setFormValues((prev) => ({ ...prev, [field]: value }));
+    setValue(field, value as CreatorRequestValues[typeof field], {
+      shouldDirty: true,
+      shouldValidate: true,
+    });
+
     setFormMessage("");
   };
 
-  const validateWorkLink = (): boolean => {
-    const workLink = formValues.workLink.trim();
-
-    try {
-      new URL(workLink);
-      return true;
-    } catch {
-      setMessageTone(FORM_MESSAGE_TONE.ERROR);
-      setFormMessage(t("authCreator.form.workLinkInvalid"));
-      return false;
-    }
-  };
-
-  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    if (!isSubmitEnabled) return;
-
-    if (!validateWorkLink()) return;
-
+  const onSubmit = async (values: CreatorRequestValues) => {
     try {
       const response = await creatorRequest.mutateAsync({
-        firstName: formValues.firstName.trim(),
-        lastName: formValues.lastName.trim(),
-        email: formValues.email.trim().toLowerCase(),
-        phone: formValues.phone.trim() || undefined,
-        cvr: formValues.cvr.trim() || undefined,
-        address: formValues.address.trim(),
-        city: formValues.city.trim(),
-        postalCode: formValues.postalCode.trim(),
-        exampleWorkLink: formValues.workLink.trim(),
-        contentDescription: formValues.contentDescription.trim(),
+        firstName: values.firstName.trim(),
+        lastName: values.lastName.trim(),
+        email: values.email.trim().toLowerCase(),
+        phone: values.phone.trim() || undefined,
+        cvr: values.cvr.trim() || undefined,
+        address: values.address.trim(),
+        city: values.city.trim(),
+        postalCode: values.postalCode.trim(),
+        exampleWorkLink: values.workLink.trim(),
+        contentDescription: values.contentDescription.trim(),
       });
 
       if (!response.success) {
-        throw new Error(response.message || t("authCreator.form.submitError"));
+        throw new Error(response.message || t("authForm.errors.submitFailed"));
       }
 
-      toast.success(response.message || t("authCreator.form.submitSuccess"));
+      const successMessage =
+        response.message || "Creator application submitted successfully";
 
-      setMessageTone(FORM_MESSAGE_TONE.SUCCESS);
-      setFormMessage(response.message || t("authCreator.form.submitSuccess"));
+      toast.success(successMessage);
+      setMessageTone("success");
+      setFormMessage(successMessage);
 
-      setFormValues(INITIAL_CREATOR_FORM);
+      reset();
       router.push(PATHS.AUTH_CREATOR_REQUEST_SENT);
     } catch (error) {
-      setMessageTone(FORM_MESSAGE_TONE.ERROR);
-      setFormMessage(
-        error instanceof Error
-          ? error.message
-          : t("authCreator.form.submitError"),
-      );
+      applyFieldErrors(error, setError);
+      setMessageTone("error");
+      setFormMessage(getErrorMessage(error, "authForm.errors.submitFailed"));
     }
   };
 
   return {
-    formValues,
-    setFormValues,
+    methods,
     formMessage,
     messageTone,
-    isSubmitEnabled,
-    isPending: creatorRequest.isPending,
+    isValid,
+    errors,
+    isSubmitting: creatorRequest.isPending,
     updateField,
-    handleSubmit,
+    handleFormSubmit: handleSubmit(onSubmit),
   };
 }
