+++ apps/web/hooks/auth/useViewerSignUpForm.ts
@@ -1,118 +1,113 @@
 "use client";
 
-import { useMemo, useState, FormEvent } from "react";
-import { useRouter } from "next/navigation";
+import { useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { normalizeApiError } from "@/lib/http/errors/apiError";
-import {
-  INITIAL_VIEWER_FORM,
-  PASSWORD_FIELD_KEYS,
-  PasswordVisibility,
-  ViewerFormValues,
-  ViewerFieldKey,
-  VIEWER_FIELDS,
-} from "@/utils/signup";
-import { useViewerSignUp } from "./useViewerSignUp";
-import { PATHS } from "@/utils/path";
+import { useRouter } from "next/navigation";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import { useApiErrorMessage } from "@/lib/http/useApiErrorMessage";
 import { persistAuthSession } from "@/lib/auth/authSession";
+import { createViewerSignupSchema } from "@/lib/validation/schema";
+import { PATHS } from "@/utils/path";
 
-export type PasswordFieldKey = keyof PasswordVisibility;
+import { PasswordVisibility, ViewerFormValues } from "@/utils/signup";
+import { useViewerSignUp } from "./useViewerSignUp";
 
 export function useViewerSignUpForm() {
-  const router = useRouter();
   const { t } = useTranslation();
-  const { mutateAsync: viewerSignUp, isPending: isSubmitting } =
-    useViewerSignUp();
-  const [formValues, setFormValues] =
-    useState<ViewerFormValues>(INITIAL_VIEWER_FORM);
+  const router = useRouter();
+
+  const { mutateAsync: viewerSignUp, isPending } = useViewerSignUp();
+  const { getErrorMessage, applyFieldErrors } = useApiErrorMessage();
+
+  const [formError, setFormError] = useState("");
   const [passwordVisibility, setPasswordVisibility] =
     useState<PasswordVisibility>({
       password: false,
       repeatPassword: false,
     });
-  const [submitted, setSubmitted] = useState(false);
-  const [formError, setFormError] = useState("");
 
-  const passwordsDoNotMatch =
-    formValues.repeatPassword.length > 0 &&
-    formValues.password !== formValues.repeatPassword;
-
-  const isSubmitEnabled = useMemo(
+  const schema = useMemo(
     () =>
-      VIEWER_FIELDS.every((field) => Boolean(formValues[field.key].trim())) &&
-      formValues.agreed &&
-      !passwordsDoNotMatch,
-    [formValues, passwordsDoNotMatch],
+      createViewerSignupSchema({
+        fullNameRequired: t("viewerSignup.form.fixHighlightedFields"),
+        emailRequired: t("authForm.errors.emailRequired"),
+        emailInvalid: t("authForm.errors.emailInvalid"),
+        passwordRequired: t("authForm.errors.passwordRequired"),
+        repeatPasswordRequired: t("viewerSignup.form.fixHighlightedFields"),
+        passwordMismatch: t("viewerSignup.form.passwordMismatch"),
+        consentRequired: t("viewerSignup.form.fixHighlightedFields"),
+      }),
+    [t],
   );
+
+  const methods = useForm<ViewerFormValues>({
+    resolver: zodResolver(schema),
+    mode: "onChange",
+    defaultValues: {
+      fullName: "",
+      email: "",
+      password: "",
+      repeatPassword: "",
+      agreed: false,
+    },
+  });
+
+  const {
+    handleSubmit,
+    setValue,
+    setError,
+    formState: { isValid, errors },
+  } = methods;
 
   const updateField = (
     field: keyof ViewerFormValues,
     value: string | boolean,
   ) => {
-    setFormValues((prev) => ({ ...prev, [field]: value }));
-    if (formError) {
-      setFormError("");
-    }
+    setValue(field, value as never, {
+      shouldDirty: true,
+      shouldValidate: true,
+    });
+
+    if (formError) setFormError("");
   };
 
-  const togglePassword = (key: keyof PasswordVisibility) => {
+  const togglePassword = (field: keyof PasswordVisibility) => {
     setPasswordVisibility((prev) => ({
       ...prev,
-      [key]: !prev[key],
+      [field]: !prev[field],
     }));
   };
 
-  const getSubmitErrorMessage = () => {
-    if (passwordsDoNotMatch) {
-      return t("viewerSignup.form.passwordMismatch");
-    }
-
-    return t("viewerSignup.form.fixHighlightedFields");
-  };
-
-  const getPayload = () => ({
-    fullName: formValues.fullName.trim(),
-    email: formValues.email.trim(),
-    password: formValues.password,
-    confirmPassword: formValues.repeatPassword,
-  });
-
-  const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setSubmitted(true);
-    if (!isSubmitEnabled) {
-      setFormError(getSubmitErrorMessage());
-      return;
-    }
+  const onSubmit = async (values: ViewerFormValues) => {
     setFormError("");
 
     try {
-      const response = await viewerSignUp(getPayload());
+      const response = await viewerSignUp({
+        fullName: values.fullName.trim(),
+        email: values.email.trim(),
+        password: values.password,
+        confirmPassword: values.repeatPassword,
+      });
 
       persistAuthSession(response);
+
       router.push(PATHS.AUTH_SIGNUP_VIEWER_PREFERENCES);
-    } catch (err) {
-      const apiError = normalizeApiError(err);
-      setFormError(apiError.message || t("viewerSignup.form.signupFailed"));
+    } catch (error) {
+      applyFieldErrors(error, setError);
+      setFormError(getErrorMessage(error, "viewerSignup.form.signupFailed"));
     }
   };
 
-  const isPasswordField = (
-    fieldKey: ViewerFieldKey,
-  ): fieldKey is PasswordFieldKey =>
-    PASSWORD_FIELD_KEYS.includes(fieldKey as PasswordFieldKey);
-
   return {
-    formValues,
-    passwordVisibility,
-    submitted,
+    methods,
+    isValid,
+    errors,
+    isSubmitting: isPending,
     formError,
-    isSubmitting,
-    isSubmitEnabled,
-    passwordsDoNotMatch,
+    passwordVisibility,
     updateField,
     togglePassword,
-    handleSubmit,
-    isPasswordField,
+    handleSubmit: handleSubmit(onSubmit),
   };
 }
